pyproject.toml
[project]
name = "cvitae-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "playwright>=1.46.0",
    "pyyaml>=6.0.3",
]
